+++ src/Folder.php
@@ -12,6 +12,7 @@
 
 namespace Webklex\PHPIMAP;
 
+use Carbon\Carbon;
 use Webklex\PHPIMAP\Exceptions\GetMessagesFailedException;
 use Webklex\PHPIMAP\Exceptions\MessageSearchValidationException;
 use Webklex\PHPIMAP\Query\WhereQuery;
@@ -426,7 +427,20 @@ class Folder {
      * @throws Exceptions\ConnectionFailedException
      */
     public function appendMessage($message, $options = null, $internal_date = null) {
-        return imap_append($this->client->getConnection(), $this->path, $message, $options, $internal_date);
+        /**
+         * Check if $internal_date is parsed. If it is null it should not be set. Otherwise the message can't be stored.
+         * If this parameter is set, it will set the INTERNALDATE on the appended message. The parameter should be a
+         * date string that conforms to the rfc2060 specifications for a date_time value or be a Carbon object.
+         */
+
+        if ($internal_date != null) {
+            if ($internal_date instanceof Carbon){
+                $internal_date = $internal_date->format('d-M-Y H:i:s O');
+            }
+            return imap_append($this->client->getConnection(), $this->path, $message, $options, $internal_date);
+        }
+
+        return imap_append($this->client->getConnection(), $this->path, $message, $options);
     }
 
     /**